Rename pathHandler to pathService, it is no more a just handler
